+++ TelegramBotApi/Application/Handlers/Urfu/GetUserInfoCommandHandler.cs
@@ -1,5 +1,4 @@
 ﻿using Application.Requests.Urfu;
-using Dal.Contracts.Interfaces;
 using Logic.Models;
 using Logic.Services.Interfaces;
 using MediatR;
@@ -8,21 +7,16 @@ namespace Application.Handlers.Urfu;
 
 public class GetUserInfoCommandHandler : IRequestHandler<GetUserInfoRequestCommand, UserInfo>
 {
-    private readonly IUrfuUserDataService userDataService;
+    private readonly IUserDataService _userDataService;
 
-    private readonly IUserInfoRepository userInfoRepository;
-
-    public GetUserInfoCommandHandler(IUrfuUserDataService userDataService, IUserInfoRepository userInfoRepository)
+    public GetUserInfoCommandHandler(IUserDataService userDataService)
     {
-        this.userDataService = userDataService;
-        this.userInfoRepository = userInfoRepository;
+        this._userDataService = userDataService;
     }
 
     public async Task<UserInfo> Handle(GetUserInfoRequestCommand request, CancellationToken cancellationToken)
     {
-        // достаем данные о пользователе по id, если их нет, то получаем новые, записываем их по userid
-
-        var userInfo = await userDataService.GetUserInfo(request.UserKey);
+        var userInfo = await _userDataService.GetUserInfo(request.UserKey);
 
         return userInfo;
     }
