+++ TelegramBotApi/Application/Handlers/Urfu/GetUserMarksCommandHandler.cs
@@ -8,13 +8,13 @@ namespace Application.Handlers.Urfu;
 
 public class GetUserMarksCommandHandler : IRequestHandler<GetUserMarksRequestCommand, IEnumerable<UserMark>>
 {
-    private readonly IUrfuUserDataService userDataService;
+    private readonly IUrfuUserServerDataService _userServerDataService;
 
     private readonly IUserMarksRepository userMarksRepository;
 
-    public GetUserMarksCommandHandler(IUrfuUserDataService userDataService, IUserMarksRepository userMarksRepository)
+    public GetUserMarksCommandHandler(IUrfuUserServerDataService userServerDataService, IUserMarksRepository userMarksRepository)
     {
-        this.userDataService = userDataService;
+        this._userServerDataService = userServerDataService;
         this.userMarksRepository = userMarksRepository;
     }
 
@@ -22,7 +22,7 @@ public class GetUserMarksCommandHandler : IRequestHandler<GetUserMarksRequestCom
     {
         // достаем данные о пользователе по id, если их нет, то получаем новые, записываем их по userid
         
-        var userMarks = await userDataService.GetUserMarks(request.UserKey);
+        var userMarks = await _userServerDataService.GetUserMarks(request.UserKey);
 
         return userMarks;
     }
